Bugfix in reference counting.
Storing windows in a selfdeleting (UAP) container would
crash. This has been overcome since the bugfix for the
containers self has been taken.

The wxWidgets controls or windows are not reference
counted, or have its own one. So they should not be stored
simply in a container of my framework.

Special reference increment is done for the specific dialogs.